Reworking image rendering
In the past, I was using some images as links. I believe this is a bad
habit. I am not providing a visual clue that the image is linkable.

This commit removes the linking of images from the short-code. I will
instead favor explicit textual links.

In addition, as I'm reviewing marginfigure, I realize that I may be
able to normalize the images to wrap all in figures.