Reinstate marginalia in the margin
But eliminate floats as we do so. Annoyingly, I can't just push them into the
`sb` grid element, but have to report to absolute positioning. I want the
first line of the sidenote to stay aligned with the point in the body copy
where I introduce the note, rather than aligning on the paragraph that
contains it, which is why it's wrapped within an object. I might also want to
revisit the use of 'unparagraphy' as a partial, but I'll cross that bridge
when I need to cross it.